requirements.txt
discord.py>=2.3.2
aiosqlite>=0.19.0
aiohttp>=3.8.0
apscheduler>=3.10.0
matplotlib>=3.7.0
numpy>=1.24.0
pandas>=2.0.0
yfinance>=0.2.0
python-dotenv>=1.0.0
flask>=2.3.0
requests>=2.31.0
